Add domain model for check tax number validity
