+++ erica/application/JobService/job_service_factory.py
@@ -4,17 +4,16 @@ from typing import Callable, Type
 from opyoid import Injector
 
 from erica.application.ApplicationModule import ApplicationModule
-from erica.application.FreischaltCode.FreischaltCode import BaseDto
 from erica.application.FreischaltCode.Jobs.jobs import activate_freischalt_code, request_freischalt_code, \
     revocate_freischalt_code
 from erica.application.JobService.job_service import JobService, JobServiceInterface
 from erica.application.tax_declaration.tax_declaration_jobs import send_est
-from erica.application.tax_number_validation.check_tax_number_dto import CheckTaxNumberDto
 from erica.domain.FreischaltCode.FreischaltCode import FreischaltCodeRevocatePayload, FreischaltCodeActivatePayload, \
     FreischaltCodeRequestPayload
 from erica.domain.Shared.BaseDomainModel import BasePayload
 from erica.domain.Shared.EricaAuftrag import RequestType
 from erica.domain.TaxDeclaration.TaxDeclaration import TaxDeclarationPayload
+from erica.domain.tax_number_validation.check_tax_number import CheckTaxNumberPayload
 from erica.erica_legacy.request_processing.requests_controller import UnlockCodeRequestController, \
     UnlockCodeActivationRequestController, UnlockCodeRevocationRequestController, EricaRequestController, \
     CheckTaxNumberRequestController, EstRequestController
@@ -59,7 +58,7 @@ def _freischalt_code_revocation_injector():
 def _check_tax_number_injector():
     module = ApplicationModule()
     module.bind(Type[EricaRequestController], to_instance=CheckTaxNumberRequestController)
-    module.bind(Type[BaseDto], to_instance=CheckTaxNumberDto)
+    module.bind(Type[BasePayload], to_instance=CheckTaxNumberPayload)
     module.bind(JobServiceInterface, to_class=JobService)
     module.bind(Callable, to_instance=revocate_freischalt_code)
