+++ erica/application/tax_number_validation/check_tax_number_dto.py
@@ -1,34 +1,7 @@
-from enum import Enum
-
 from erica.application.base_dto import BaseDto
-
-
-class StateAbbreviation(str, Enum):
-    bw = "bw"
-    by = "by"
-    be = "be"
-    bb = "bb"
-    hb = "hb"
-    hh = "hh"
-    he = "he"
-    mv = "mv"
-    nd = "nd"
-    nw = "nw"
-    rp = "rp"
-    sl = "sl"
-    sn = "sn"
-    st = "st"
-    sh = "sh"
-    th = "th"
-
-    # To find the correct values case insensitively
-    @classmethod
-    def _missing_(cls, value):
-        for member in cls:
-            if member.lower() == value.lower():
-                return member
+from erica.domain.tax_number_validation.check_tax_number import CheckTaxNumberPayload
 
 
 class CheckTaxNumberDto(BaseDto):
-    state_abbreviation: StateAbbreviation
-    tax_number: str
+    ttlInMinutes: int
+    payload: CheckTaxNumberPayload
