+++ tests/application/job_service/test_job_service_factory.py
@@ -2,11 +2,11 @@ import pytest
 
 from erica.application.FreischaltCode.FreischaltCode import BaseDto
 from erica.application.JobService.job_service_factory import get_job_service
-from erica.application.tax_number_validation.check_tax_number_dto import CheckTaxNumberDto
 from erica.domain.FreischaltCode.FreischaltCode import FreischaltCodeRequestPayload, FreischaltCodeActivatePayload, \
     FreischaltCodeRevocatePayload
 from erica.domain.Shared.EricaAuftrag import RequestType
 from erica.domain.TaxDeclaration.TaxDeclaration import TaxDeclarationPayload
+from erica.domain.tax_number_validation.check_tax_number import CheckTaxNumberPayload
 from erica.erica_legacy.request_processing.requests_controller import UnlockCodeRevocationRequestController, \
     UnlockCodeRequestController, CheckTaxNumberRequestController, UnlockCodeActivationRequestController, \
     EstRequestController
@@ -52,7 +52,7 @@ class TestJobServiceFactory:
 
         assert isinstance(job_service.repository, EricaRequestRepository)
         assert isinstance(job_service.background_worker, BackgroundJobRq)
-        assert issubclass(job_service.payload_type, CheckTaxNumberDto)
+        assert issubclass(job_service.payload_type, CheckTaxNumberPayload)
         assert issubclass(job_service.request_controller, CheckTaxNumberRequestController)
 
     def test_if_send_est_type_then_return_correctly_configured_service(self):
