+++ tests/erica_legacy/test_routes.py
@@ -11,7 +11,7 @@ from erica.erica_legacy.api.v1.endpoints.tax import is_valid_tax_number, get_tax
 from erica.erica_legacy.api.v1.endpoints.unlock_code import request_unlock_code, activate_unlock_code, revoke_unlock_code
 from erica.erica_legacy.pyeric.eric import EricResponse
 from erica.erica_legacy.pyeric.pyeric_controller import GetTaxOfficesPyericController
-from erica.application.tax_number_validation.check_tax_number_dto import StateAbbreviation
+from erica.domain.tax_number_validation.check_tax_number import StateAbbreviation
 from tests.erica_legacy.samples.grundsteuer_sample_data import get_grundsteuer_sample_data
 
 from tests.erica_legacy.utils import create_unlock_request, create_unlock_activation, create_est, create_unlock_revocation, \
